python-lambda/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chat"
version = "0.1.0"
dependencies = [
    "google-genai"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
